iam_role_eks/requirements.txt
boto==2.49.0
boto3==1.26.155
kubernetes==22.6.0
PyYAML==6.0.1